llm/information_shaping/requirements.txt
langchain
pdfminer.six
unstructured[local-inference]
openai
tiktoken
faiss-cpu
gradio